# api/src/entities/User.entity.ts
import { Column, CreateDateColumn, Entity, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm'

@Entity('users')
export class User {
  @PrimaryGeneratedColumn('uuid')
  id!: string

  @Column({ type: 'citext', unique: true })
  email!: string

  @Column({ type: 'text' })
  password_hash!: string

  @Column({ type: 'varchar', length: 100, nullable: true })
  display_name!: string | null

  @Column({ type: 'text', nullable: true })
  avatar_url!: string | null

  @Column({ type: 'varchar', length: 16, default: 'user' })
  role!: 'user' | 'admin'

  @Column({ type: 'varchar', length: 16, default: 'active' })
  status!: 'active' | 'suspended' | 'deleted'

  @CreateDateColumn({ type: 'timestamptz' })
  created_at!: Date

  @UpdateDateColumn({ type: 'timestamptz' })
  updated_at!: Date
} 